backend/requirements.txt
# Core
fastapi==0.115.*
uvicorn[standard]==0.34.*
python-dotenv==1.1.*
pydantic==2.10.*
pydantic-settings==2.7.*

# Database
sqlalchemy[asyncio]==2.0.*
asyncpg==0.30.*
psycopg2-binary==2.9.*
alembic==1.14.*

# Redis
redis==5.2.*

# Auth
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
python-multipart==0.0.*

# HTTP
httpx==0.28.*

# Utils
unidecode==1.3.*
python-slugify==8.0.*

# Geolocation
geopy==2.4.*

# AI (Google Gemini)
google-genai==1.*
